Created utils.py to hold functions like constant_time_compare and base64-encoding
